api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
tensorflow-cpu==2.19.0
numpy==2.0.2
pillow==11.1.0
